Don't try to destroy an element that's already destroying

+++ packages/ember-views/lib/views/view.js
@@ -1674,6 +1674,8 @@ Ember.View = Ember.CoreView.extend(
     @return {Ember.View} receiver
   */
   destroyElement: function() {
+    if(this.isDestroyed || this.isDestroying) { return; }
+
     return this.currentState.destroyElement(this);
   },
